needs_refcount now a property of CompiledType subclasses
Before needs_refcount was determined by the parser, but it is really
more the property of the type. This commit makes needs_refcount be a
property of CompiledType subclasses, while keeping a shortcut on Parser.

needs_refcount is now a property, not a function call, so client code
had to be adapted.

* parsers.py:

  (CompiledType): Enforce implementation of needs_refcount.
  (BasicType): Put needs_refcount to false for basic types.
  (ASTNode): needs_refcount == True for ASTNodes
  (Parser): shortcut to needs_refcount, returns False when get_type ==
            None.
  (Or, Tok, TokClass, etc.): Remove needs_refcount method

* templates: Adapt to the fact that needs_refcount is now a property

N918-023

Change-Id: Ic2b54a78748501b3886ddf89ab86a2c792286aa2