First version of the JSON pretty-printer
Addition of a JSON pretty-printer to libadalang. This can be useful to
leverage visualisation/analysis tools that work on JSON data in general.
More coming about this in the future.

The JSON pretty printer is based on boost::property_tree, which has a
built-in JSON pretty printer. This can be useful because we can leverage
the same technique to produce XML trees in the future if needed.

compiler/templates/cpp:

* astnode_type_[impl|def].mako: Implement get_property_tree method on
  ASTNode subclasses.
* enum_type_[impl|decl].mako: Implement get_ptree static function for
  enums.
* interactive_main.mako: Print json if the -j flag is passed

support:

* ast.hpp: Add the abstract method definition to the root AST
  class.

  Add implementation of get_property_tree on lists.

* context.cpp: Add print_json helper method on AnalysisUnit.

* lexer.hpp: Add get_property_tree on SourceLocations and
  SourceLocationRanges

Change-Id: I2567b9f3fe2d5a78a31f373fa3c680b81bfb4fbd