Make the name of the test binary static and the main header
Remove the parameter in the compile context that used to enable one to
tune the name of the test binary ("parse") and of the main header. It's
currently always has the same value and we don't know what it could be
used for.

This simplifies #include directives in templates.

Preparatory work for O310-018

Change-Id: I10d667c3eed8024a78fa1edd11d2e42dc83519ea