O323-021: Add an error mode to the Opt parser.
This approach is a simpler alternative to the approach described in
NA30-012. The idea is that you can flag an Opt parser to emit an error
when the sub parser is not found, but still continue parsing normally.

This commit implements this approach and deploys it for semicolon in
statements, as well as for closing parenthesis in subprogram
specifications, because it has been determined that those cases were non
ambiguous.

Detailled changelog:

* bodies.py: Make ";" recoverable in statement lists.
* decl.py: Make ")" recoverable in subprogram parameter lists end.

* parsers.py (is_tok): Primitive to check if a parser is a token parser.
  Add it to render_template's default env.

  (Opt.error): Transform the current parser into a recovery parser.

* opt_code.mako: Implement the error recovery: When the token is not
  found, emit a diagnostic.

Change-Id: Ifad92a0e05103b6ed5ed4db155564041c74932c7