Testsuite: make it possible to debug a program without running memcheck
For some programs in the future (for instance, Python interpreters), we
might want not to run Valgrind's memory checking capabilities but still
to debug them when running the testsuite in debug mode. Add the
necessary bits to differenciate these two features.

Also make it possible to run Valgrind multiple times in the same
testcase (reports are merged).

Preparatory work for O323-002

Change-Id: I11bff51c1168fa2826ccab1311c67a8c3192b53d