Refactor the indentation engine and put it in a dedicated unit
The identation engine does not need to be generated. Besides, having it
in a dedicated unit makes it easier to debug it (yay, static line
numbers!). The refactoring is done as an attempt to finally understand
what the code actually does, and then to review patches on it.

Followup on O313-038

Change-Id: Ia594b79df91f821a5713485ec5735919b65d6aac