Simplify the indentation algorithm to rely on lookups
The previous algorimth used to visit all AST nodes and sometimes lookup
nodes for the beginning of lines. This change makes it just iterate on
every line, query the sloc for the node that starts the line and use its
indentation level.

This change fixes indentation for the trailing lines on which no AST
node start their source location.

Followup on O313-038

Change-Id: I74cd8ac6eb9ca89e7001e4619e6a53cdb75454d0