C/Python API: expose tokens as AST node fields
With very basic support, essentially just getting the text of the token.
Getting the kind of tokens, or the whole stream from the AnalysisUnit,
is not supported yet. Only token properties on ASTNodes are bound.

* parsers.py:
  (is_token_type): New predicate.
  (make_renderer): Make is_token_type available in templates. Also make
  the "token" CAPIType available.
  (Token): Add c_type and py_type for Token type.
  (ASTNode.get_public_fields): Return all fields including token fields.

* c_astnode_field_access_impl.mako: Handle token fields.

* c_header.mako: Declare a data type for tokens. Add "token_text"
  as a token primitive.

* c_body.mako: Implement the "token_text" primitive.

* ast_subclass.mako: Handle token fields.

* module.mako: Add a Token class, bind the token C type and the
  "token_text" primitive.

Also update "field_access" tests for C and Python APIs to cover token
fields.

Followup on O404-004

Change-Id: I8af222e92596a317d39106c8505f990b961264a3