Compiler: setup infrastructure to correctly handle casing conventions
Part of O413-016

Depending on the target language (C++, C, Python) and the usage of
identifiers (class name, enumerator, function name, etc.), we want use
different casing conventions. Until now, this was done in a very ad hoc
way across the compiler.

This commit introduces a helper to provide centralized identifier
formatting utilities and updates all parts of the compiler that must
deal with casing conventions. In particular, it fixes the casing of
ASTNode subclasses and enumerations in the C API. In order to properly
accomplish this goal, this commits also updated the Ada parser
specification in order to specify accurately the structure of
identifiers.

As it is a very adjacent issue, it also adds keywords clashing handling
so that language specification writers do not have to think about what
identifiers will be valid during code generation. This introduces a tiny
change of API in Python for enumerators.

Change-Id: I497220d8e68ae1fc640198ac623bae7b151792a3