Switch file lexing to GNATCOLL.Mmap to workaround Quex
Fix for O402-026

Quex's file-based lexing makes the process exit when the file cannot be
open. Workaround this reading ourselves files into a buffer and using
Quex's buffer-based lexing.

In case of error, also propagate an exception the Ada code, return a
NULL pointer in the C API and raise an IOError in the Python API.

Change-Id: I6281d34c1a5b20e593a966e77d6f535d8cb0f13e