First pass on the static extensions system.
The behavior of this extensions system is entirely file based: We add
hooks in the regular templates to include specific extensions files.

At the root of a specific langkit language definition, the engine will
go looking in the "extensions" subfolder. If an extension is found, it
will be included, else, nothing will be done.

This commit adds the base mechanism, as well as a a first extension
point, allowing the user to add fields to ast nodes subclasses.

* compile_context.py: Add extensions_dir property, that will be set by
  libmanage to be $lang_root/extensions.

  (ext): Helper function that will return a specific extension file
  given a series of names or strings.

* template_utils.py: Pass a TemplateLookup at template instantiation
  time, to correctly resolve templates in include/namespace directives.

* extensions.mako: Helper def block to include an extension only if it
  exists

* astnode_type_def_ada.mako: Add an extension point,
  nodes/$type_name/components, to add components to node types

OA27-037

Change-Id: I30fb689d409debd76c9037d78d64d294626a853f