Add charset selection to public APIs and update Ada material
This change makes it possible for the users of Langkit-generated
libraries to select the charset to use to decode input sources. As a
consequence, it also updates the Ada testsuite to the new APIs and
introduce a couple of testcases to check that charset handling is done
properly.

There are still two things left to do in order to have the Unicode
support in Langkit done API-wise:

  * detect early when the provided charset is not supported and raise a
    well-specified error in this case;
  * specify what happens when the input is badly encoded (for instance
    when decoding ISO-8859-1 sources as UTF-8 ones) and implement it:
    currently the Quex lexer just calls exit(), which is not acceptable.

Change-Id: Ifc76427143ffcec9fe56832b16d1372eaf0463c9
TN: O414-018