Do not use designated_env in the scope machinery
As designated_env will soon yield orphan lexical environments, it is not
suitable to build the lexical environment in the populate pass. Indeed,
this pass needs lexical environment with regular parents: it's only once
that we use populated envs that orphan ones are useful.

This change adjusts the scope machinery not to rely on designated_env.
It introduces the "env_for_scope" properties to be used instead in this
context.

Change-Id: I8b275fd1051c4411de654fa6d928f6fca7cf5703
TN: P415-027