Testsuite: give more control to spawned subprocess output redirection
In particular, make it possible for test drivers to get the programs
output as a returned string and make it possible not to append this
output to the testcase output file.

This is needed to check consistent output for the Python/Ada symbol
resolution drivers without cluttering testcase output.

Change-Id: I52f9f4148bb8039fdd3e4758ef35d33c0a078aaf
TN: P509-043