P919-020 public-api: Convert every boolean field to qualifier nodes
A new notion of Qualifier node was introduced in langkit. It is meant to
store a boolean result, but in a regular AST node, and with the added
information that a node contains.

Those qualifier nodes can be converted to booleans easily via the
As_Bool property. Here is an example using the python API::

    >>> u = ctx.get_from_file("foo.adb")
    >>> print u.root.find(PrivateQualifier).p_as_bool
    False

You can also get the sloc they would be at in the source if they were
present::

    >>> print u.root.find(PrivateQualifier).sloc_range.start
    <Sloc 2:1 at 0x7fc3916ed7d0>

end-public-api

Change-Id: Ib14cffc0c6f4beaff165b1d0b1a9a1071b0c4e98