P919-023 public-api: Refactor TypeDiscriminant
Before, the discriminant part of a type was stored in a
`TypeDiscriminant` node, and and unknown discriminant part was
represented as a `TypeDiscriminant` node with an empty list of
discriminants, which is incorrect.

Now, `TypeDiscriminant` has been renamed into `DiscriminantPart`. A list
of known discriminants will result in the parsing of a
`KnownDiscriminantPart` node containing a list of discriminants. An
unknown discriminant part will result into the parsing of an
`UnknownDiscriminantPart` node containing nothing.

.. code-block:: shell

    % parse -r discriminant_part "(A : Integer)"
    KnownDiscriminantPart[1:1-1:14]
    | discr_specs:
    | | DiscriminantSpec[1:2-1:13]
    | | | ids:
    | | | | Id[1:2-1:3]
    | | | | | tok: A
    | | | type_expr:
    | | | | SubtypeIndication[1:6-1:13]
    | | | | | has_not_null:
    | | | | | | NotNullQUalifierAbsent[1:6-1:6]
    | | | | | name:
    | | | | | | Id[1:6-1:13]
    | | | | | | | tok: Integer

    % parse -r discriminant_part "(<>)"
    UnknownDiscriminantPart[1:1-1:5]

end-public-api

Change-Id: If1efa3b9b07f0ba413d0308704f2fa5dcd24b99f