manage.py: fix --build-dir's handling for performance testing
The --build-dir option is used to customize the name of the directory
used to build Libadalang for perf testing. However, it is always the
default one ("build") that is currently used to run testing, so it's the
same LAL that is tested in all cases. Fix it to use the one that is
built with --build-dir.

Change-Id: I889c8e15e259025acaf0df44458dd5dc0bf0b652
TN: minor