Project_Unit_File_Provider: optionally release the input project
This adds an argument to the constructor of
Project_Unit_File_Provider_Type objects so that the project it wraps can
be automatically free'd when the unit file provider is free'd itself.
This will be useful in the C API, as users will not see the project
itself.

Change-Id: I54ce2667ce622f3a3f5b6ff2bc6b3c8d4640aad1
TN: PA05-041