Testsuite: make it possible for Python testcase to choose their interp.
This is either the default ("python" or the Python interpreter specified
with --with-python=) or the one used to run the testsuite ("self").

This will make it possible for testcases like stylechecks to use the
same interpreter as the testsuite to be able to use PEP8 and other
modules.

Change-Id: I2a45c2cfd8f8b5c2533b8bf6871659d92e38dfa6
TN: Q114-003