P506-012: Refactor labels/blocks xrefs
Introduce a new class, NamedStmt, to wrap named statements, and
introduce a BasicDecl as a subcomponent, so we respect both the
statements hierarchy and the invariant that entities stored in envs are
BasicDecl.

Refactor lexer/parser so that "<<" and ">>" are tokens, and the label is
just an identifier.

Remove specific resolution and use Name resolution for labels/blocks.

Change-Id: I196c66a1af5f21b6526fd63aefc518f4172cef23