Move install/script Travis rules to dedicated shell scripts
This makes it very easy to reproduce locally errors that occurs
with Travis.

As a bonus, the change also make the install/build stop as soon as there
is one error. Most of the time, all the commands that follow will fail
and flood the output, making it hard to see what is the first thing that
went wrong.

Change-Id: Id88c2cd16399283b99473ec50f01c8299a386ef6
TN: minor