P506-012: Massive refactoring of type binding
Before, we tried to use the canonical type everywhere in equations.
Now we use the subtype/type view available at the point, and let
canonicalization to type comparison, inside of type binds and of type
matches.

The result is that precise subtypes/views are propagated by equations.
This fixes a lot of bugs in name resolution.

Change-Id: Id11f13c38ecc1cd5174f860f770551e8520cc26f