P506-012: Fix&refactor array indices exprs xref
For the moment it was presumed that if an index is a range, then it is
automatically an Integer. However, it is possible for a range to be of
another type if one if its components is of another type. The type of
the range is then infered by the compiler.

Change-Id: I1fe9c22c255ee4103b7de38fdd6cf3505450dcfe