tests/python/node_identity: remove testcase
The point of this testcase is to check that AST node wrapper instances
in the Python API are unique for each AST node. However, with the
incoming switch to entities in the whole API, this feature no longer
makes sense: one single node will have various wrappers, as they will
hold different entity information.

TN: Q920-012