RB22-027: Fix Iterator_Element aspect referring to incomplete type.
Previous implementation was calling name_designated_type, which rebinds
origin to Self. This is wrong, as an Ada 2012 Iterator_Element aspect can
refer to an incomplete type and therefore Self may not have the complete
view on the type of the iterated elements. Instead, we call
name_designated_type_impl directly without rebinding origin so as to preserve
the view from the caller.