P506-012: Fix designated_type on fully qualified names.
This fixes an infinite recursion when trying to find the designated type
of a fully qualified name: while computing designated_type on a dotted name,
the generic designated_env is called on the prefix, which includes primitive
envs. But to compute the primitive envs, designated type is called again on type
declarations appearing in the env of the prefix, causing an infinite recursion.

This commit addds designated_env_no_overloading which can be used when we know
that the name to resolve is an entity that cannot be overloaded (e.g. a type),
and that does require looking in primitive environments.