c_api/field_access: remove a pointless check
Because of field alignments, there are uninitialized fields in the C
structure for entities, so it is invalid to use memcmp to compare them.
This check was pointless anyway: we don't necessarily want to ensure
that a function does not clobber its OUT arguments when it returns an
error.

TN: S627-006