S528-024: Improve scheduling of populate_lexical_env on with'ed units.
Until now, a with'ed unit would be populated as soon as the with clause was
seen, meaning the library item living in the with'ing unit would not exist yet.
This was problematic in the case of child units, because we would try to add
elements to the env of the parent package while populating a child package
triggered by a limited with clause, but this env didn't exist.

The fix consists in delaying the populate_lexical_env of with'ed units
until the library item defined in the unit is seen.