S902-009: Fix next_part.
Minimal lookups are sufficient here because the next part should be
in the same environment. This also prevents finding back the original
node when doing an env lookup on the package's body (since a package
declaration is the parent of its body).

An alternative could have been to filter out the results that are
equivalent to Entity, but it seems cleaner that way.