S917-019: Get rid of all env lookups in p_next_part.
Next_part being used extensively during populate_lexical_env, it is not
the first time that we have issues with next_part because it relies on
env lookups and those are not yet fully populated.

This commit gets rid of all remaining env lookups in BaseTypeDecl's
p_next_part and replaces them by manual searches in the list of
declarations.