S917-041: Don't implement name_designated_type on QualExpr.
This causes issues when one wants to use name_designated_type to know
if a given name refers to a type, such as in AttributeRef's image_equation.
Indeed, a QualExpr is an expression and not a type expression. Note that
It's still possible to retrieve the type denoted by the prefix using
"designated_type".