Testsuite: propagate the build mode to gprbuild calls
Testcases for the C and Ada API call gprbuild in order to build test
programs. This works fine when Libadalang was installed using GPRbuild,
but yields glitches when Libadalang comes from the build directory, as
it still has the BUILD_MODE scenario variable that defaults to "dev": if
the Libadalang tested was actually built in "prod" mode, then each test
will try to recompile Libadalang concurrently, leading to weird link
errors.

This commit propagates the build mode from "ada/manage.py test" to the
gprbuild calls in testcases to avoid this.

TN: SA30-026