R219-008: Fix erroneous refactoring.
While removing the p_unit_body property as part of the
"multiple units per file" patch, I made an error while copying its
logic in the p_basic_decl_next_part_for_decl property by discarding
the case where the enclosing compilation unit is a Body, in which
the "body unit" is simply itself.

This was causing a weird error in the ALS because it ended up fetching
the "body unit" in another file.