LAL.Project_Provider: ease the use of aggregate projects
As a shortcut to the use of Create_Project_Unit_Providers (i.e.
returning a list of unit providers), make Create_Project_Unit_Provider
(i.e. returning only one unit provider) run the project partitionner
when no sub-project is provided and fail if and only if it needs to
create multiple unit providers.

Previously, Create_Project_Unit_Provider used to consider the root
project as the sub-project to use, which did not work for most aggregate
projects.

TN: SA09-018