Add a missing use of GPR lock in the project unit provider C constructor
The only known way to reproduce for now is to use the C API or the OCaml
(Python has a global lock, the Ada API is not concerned) one and to run
several threads to analyze big projects to have a chance to trigger the
race condition in the project loading. Hence, we do not have right now
reasonable ways to trigger this bug in our testsuite: not adding a
testcase for now.

TN: SC11-018