nameres.adb: refactor the processing of pragmas
This patch splits the decoding of pragmas and the execution of the
corresponding actions in two separate subprograms for clarity: it is now
very easy to see the list of supported pragmas and their expected
syntax. Besides, pragmas processing code is now less nested.

As a bonus, pragmas decoding is more strict. For instance, unexpected
names in arguments are now always rejected.

TN: SC13-001