T103-024: Rework stop_resolution mechanism.
The resolve_own_names property is introduced, which performs name resolution
on a node up to stop_resolution boundaries, that is, without recursing on
stop_resolution nodes.

As a consequence, nodes that have stop_resolution set are able to call
p_referenced_decl on a sibling without triggering an infinite loop,
which will be useful for the next commits.