T103-024: Rework resolution of array attributes' arguments.
AttributeRefs such as 'Range, 'First & co. that accept static argument
are now considered entry points: indeed, their resolution does not
depend on the surroudning context.

As a consequence, this allows the static expression evaluator to call
p_referenced_decl on any name in the arguments of the attribute without
the need of the try_immediate option, because the infinite recursion was
removed thanks to the added entry point.