T103-024: Rework resolution of aggregates.
The construction of the xref equations needed to resolve Aggregate
assocations were moved to AggregateAssoc itself instead of being done in
BaseAggregate, which means that each association may now be resolved
separately.

This is very important, because it means we can now resolve discriminant
association independently of the other associations.

Thanks to the change on the stop_resolution mechanism and on the fact that
all aggregate associations are now in different resolution trees, resolution
of a component association of a variant record aggregate can now safely call
p_referenced_decl on one of its discriminant association sibling without
triggering an infinite recursion.