T706-023: Fix requeue stmt equation.
An assumption was previously made during the construction of requeue
statements' equations that the requeued name should reference an EntryDecl,
thus filtering off entry bodies.

This was not a problem until recent improvements made on entry navigation
(done under this ticket), where entry bodies were now added to lexical
envs and became thus visible. This fix consists in not binding twice the
requeued name.

The simple change in the test consists in moving the body of the
requeued entry such that it is visible at the site of the requeue stmt.
This shows that even though the entry body is visible, the requeue stmt
is correctly resolved.