TC07-049 Update GitHub Action to build Linux/Windows
binaries and drop appveyor. The common script is used to
build Linux, Windows and MacOS with debuf and production
modes, as static and relocatable. Currently for ALS we
use only a relocatable/debug, but this doesn't work any
more, so I'm going to rewrite ALS to use production build.

Drop uploading from main.yml, because we don't need it any
more.