U219-050 fix memory corruption in OCaml binding for auto_provider
The C strings allocated to represent the list of files in the project
were garbage-collected too aggressively by the GC, resulting in memory
corruption. Switch to a version of the code where we keep more GC roots.

Also explicitly set the last element of the array passed to LAL
to NULL. It is not clear this was done implicitly by Ctypes.

(cherry picked from commit 90963281ea3425b9b68c0e8dcedc1881e4e74ff1)