U219-050 fix more memory corruption problems with UnitProvider binding
Based on what we found in Ctypes discussions [1, 2, 3], and what we
observed in practice using Valgrind, the OCaml lifetime for the C
strings containing the names of the files in the project for an auto
provider is too short. The garbage collector can collect those strings
before the function create_auto_provider has fully ran.

We correct this by forcing the cstrings local variable to be live accross
the entire call, and even further: it won't be collected until the
finalizer for the the UnitProvider runs. In practice, this eliminates
all the problems we've been able to see with Valgrind.

[1] https://github.com/ocamllabs/ocaml-ctypes/issues/476
[2] https://github.com/ocamllabs/ocaml-ctypes/issues/571
[3] http://lists.ocaml.org/pipermail/ctypes/2018-March/000276.html