U918-003: Fix name resolution of implicit `/=` operator.
This is done by reworking how implicit `/=` operators are handled. In particular,
we remove the special casing done it operator name resolution and instead rely and
lexical environments only, by adding an entry for the `/=` operator whenever we
have to add an entry for a user-defined `=` operator.

This is therefore a nice cleanup.