UA25-009: Fix regression on invalid code.
My changes for TA29-030 made LAL crash with an Assertion_Error (or worst in
production builds) when given the code pattern presented in the test case attached
to this commit: a parameter's type has the same name as the parameter itself. Note
that this is not valid Ada, however LAL shouldn't choke on it! This commit
resolves the issue.