Improve completion for `SubtypeIndication`
This patch improves the completion of `SubtypeIndication` nodes by
adding a new field in the `CompletionItem` struct: weight. This field
allows to apply a weight to a completion item in order to sort them
later on the IDE side. By doing so, the higher the weight, the more
likely is the completion.

TN: UA05-032
Ref: #908