T218-019: Use minimal lookup kind in designated_type.
Using a flat lookup recurses in the parent environments, which is clearly
invalid when trying to resolve a type for the suffix of a dotted name.
The attached testcase exhibits the problem: Invoking `designated_type` on
`Bar (12).Foo` was returning the type declaration `Foo` because the
environment for `Bar (12)` returned the defining environment of the type
`Foo`.