Fix misplaced Cut parser in param_assoc rule
Since ``Cut`` parsers are being allowed in the ``Opt`` parser soon,
the behavior of a ``Cut`` parser in an ``Opt`` will drastically
change. If this was done on purpose, the undefined behavior exploited
here won't work anymore but generate tons of parsing
errors. Nevertheless, moving out the ``Cut`` parser already improves
parsing of generic instantiation (which will be improved further with
the coming "``Cut`` in ``Opt``" parser support).

TN: V331-006