Replaces uses of GNATCOLL.Locks with GNAT.Task_Lock
GNATCOLL.Locks pulls libgnarl, which is undesirable for some Libadalang
users. GNAT.Task_Lock is a good enough replacement for that: it does not
pull libgnarl, and the fact that this lock is global is not a problem as
it is re-entrant and our only GNATCOLL.Locks-powered lock (GPR_Lock) was
global, too.

TN: U907-017