Revert "Fix misplaced Cut parser in param_assoc rule"
This reverts commit 4769e61bbfc3717fac2b1ad29a010e3552ae3a11 because
of regressions in the ada language server testsuite.

This changes are intended to improve the on coming work on the parser
regarding the ability to add cuts in optional parts (langkit!685). By
doing so, the behavior that was exploited here no longer works, still
it is required to preserve the same level of error recovery support.

TN: V606-017