Improve concatenation operator handling
This change aims to fix the handling of very long strings
concatenation which often occurs in practice. Indeed, a few hundreds
of concatenations led to a stack overflow when libadalang resolves
names. This was due to the fact that concatenation expressions, since
`&` is a binary operator, was represented as a tree. This tree being
as deep as the length of the expression, requiring deep recursions,
leading to overflows. Libadalang now handles such expression as a list
of operands to avoid those deep recursions.

TN: T526-004
Ref: #930