Preprocessing: make it possible to collect compiler opts. from a subprj.
In particular for aggregate projects, it is sometimes necessary to
restrict the analysis to a sub-project in a GPR project tree. Similarly
to the project-based unit provider, add a Project/Project_Name argument
to the subprograms that collect preprocessor-related compilation options
in project files to restrict their analysis to a sub-project.

TN: V708-022