VB22-027: Fix various ambiguities in xref equations.
These ambiguous equations (i.e. with several solutions) were previously ordered
in a specific way so that the old solver would get the correct answer.
Instead of tweaking the ordering so that the new solver now gets the correct
answer, this commit rather removes these ambiguities altogether.

The test diff is a "lucky" change: the new output is better, but it was not
expected.