Testsuite: stop rely on Valgrind's XML report
For some reason, Valgrind emits an invalid XML report in some cases, and
so our testsuite code fails to parse it. Since we get very little value
doing that, disable the XML report and switch back to the simple text
output with the "-q" option: consider any non-empty Valgrind output as
an error.

Also rework the prep/prep_data_files testcase so that it does not picks
the "valgrind-report.txt" file as a data file for the test.

TN: VC05-009