W126-014: Adjust record components' visibility in their inner nodes.
In Ada, the type expression or the default expression of a component should
not have visibility on the component itself, meaning a reference with the
name of the component will never resolve to it. This was previously not
handled properly in LAL and could cause infinite recursions during name
resolution.