Improve character enumerations expression evaluation
This change improves the static expression evaluation by treating the
`Character` (as well as the `Wide_Character` and the
`Wide_Wide_Character` variants) enumaration values as simple
integers. This is due to to way libadalang implements the `__standard`
package, where the character enumerations are not completly
represented because of their huge number.

The static evaluator can now properly handle `'Val`, `'First`, and
`'Last` attributes.

TN: W130-020