W104-020: Fix infinite recursion with case exprs and extension aggregates.
The infinite recursion occurred because we called p_expression_type on a sub-expression
of Self while constructing xref equations for Self. The trick here was to realize that
since we are in a stop_resolution case, the parent aggregate already has its type set,
and therefore we can go fetch the value of the type variable for the ancestor expression
directly (and this is exactly the purpose of stop_equation).