Libadalang.Helpers: add a --gpr2 option for App
This new internal option is meant to be used in Libadalang's testsuite
temporarily and removed once App switches to GPR2 by default. It will
allow to test the support of name resolution with sources that contain
multiple compilation units: GPR2 is able to give precise CU location
information in these configurations while GNATCOLL.Projects does not.