properties/unit_dependencies: transition main from Python to Ada
This testcase's sources contain multiple compilation units.
GNATCOLL.Projects does not handle such projects correctly, while GPR2
does. This is not a problem today, but upcoming changes to correctly
handle multi-CU sources in Libadalang will uncover the lack of support
for such projects in GNATCOLL.Projects.

Transition the main for this testcase from Python (which currently has
to use GNATCOLL.Projects) to Ada (which can use GPR2) to avoid this
issue.