properties/python_doc_name: fix Ada sources layout
This testcase was using a multi-CU-per-source-file setup without a
project file that correctly describes it. Since using such a setup is
irrelevant to what the testcase exercizes, switch to a more conventional
source layout. This fixes the current spurious "File pkg.ads not found"
warning message and this will avoid new issues with the upcoming work on
multi-CU-per-source-file setups.