Fix support for sources that contain multiple compilation units
This commit turns on the "PLE root" Langkit mechanism to be able to run
PLE on each compilation unit independently, even if compilation units
are stored in the same source file.

In addition, it extends all the unit provider where this is possible
(Auto_Provider, Project_Provider for GPR2) so that resolving a unit
yields a filename *plus* the compilation unit index in that source file.

Finally, it adds several testcases to exercize these features.