Preprocessor: fix file reader when source file cannot be read
Nowadays, file readers are supposed to return a source buffer even if
they cannot perform their jobs (for instance if an error occurs before
they can even fetch source files).

The file reader implementation for preprocessing was not updated to
follow this protocol, which caused crashes when such errors occurred.