prep/errors: fix absolute paths handling issues on Windows
Absolute paths on Windows can be hard to predict (C:\foo, C:/foo,
c:\foo, ...) and thus to remove in a robust way. Instead of trying to
post-process them, do not include them in baselines in the first place.

Also remove the test driver mechanism we introduced just for that case.