Change ordering of concat equations for better performance.
Due to eng/libadalang/langkit#725, the order of the equations in presence of
big chain of dependencies does matter for performance, and in this case the type is
propagated through each node all the way from the expected type of the assignment.
This means that if a predicate failure occurs first on a deeply nested node then
a big contradiction will be generated, whereas if it first occurs on top-level
node, a small contradiction will be generated.