Fix selected components resolution for subprograms
As defined by the ARM in 4.1.3-9.2/3:

The prefix (after any implicit dereference) shall resolve to denote an
object or value of a specific tagged type T or class-wide type
T'Class. The selector_name shall resolve to denote a view of a
subprogram declared immediately within the declarative region in which
an ancestor of the type T is declared.

That means that subprograms can be called with dot notation as long as
the first parameter is of the type of the object referenced by the
prefix, weither the given subprogram is a primitive or not.