Fix NoTypeObjectRenamingDecl name resolution (Ada 2022)
Since NoTypeObjectRenamingDecl have no subtype mark, the
ObjectDecl.expr_type and ObjectDecl.defining_env properties can't be
used because they are based on the value of the ObjectDecl.type_expr
field (which is null for NoTypeObjectRenamingDecl).

Instead, expr_type and defining_env have been overloaded to fulfill
NoTypeObjectRenamingDecl constraints.

This is an Ada 2022 feature, see AI12-0275.