Preprocessing: fix import of compiler switches for specific sources
Due to confusion between fullnames and basenames for source files in the
GNATCOLL.Projects API, the "for Switches ("<source-file>") use ..." GPR
clauses used to be incorrectly ignored by the
Libadalang.Preprocessing.Extract_Preprocessor_Data_From_Project
function.

This changes fixes this bug, which was actually already visible in the
testsuite, hence the baseline update. Note that while the GPR2
implementation was correct on this point, the testcase forgot to call
the Update_Sources procedure, and so the GPR2 implementation was not
able to take sources-specific compiler switches into account, which
explains why the bug was not visible even with the (incomplete) GPR2
circuitry in this testcase. Note that the GPR2-next API will make it
possible to detect such incomplete setups in the future.