Consider all parts of entity when looking for an aspect
The `P_Get_Aspect` property now considers all the parts of an entity
when looking for an aspect. Also, a new boolean parameter has been
introduced (`Previous_Parts_Only`) to limit the search to the previous
parts of the entity in order to comply with visibility rules (False by
default).