Refactor is_spark_impl to handle explicit `SPARK_Mode => Off` correctly
`is_spark_impl` has been refactored into `spark_mode_aspect`, which
rather than returning just a boolean for "On/Off", returns an aspect,
which contains the information "On/Off/no information".

This in turns allow to implement the logic "either take the value from
the aspect, and if there is none, take the value from the .adc".