Improve the quality of nameres tests
This change introduces a new argument that can be used for the nameres
`Test_Statement` and `Test_Statement_UID` pragmas. This boolean
argument, named `Expect_Fail` can be set to check that a given
statement is failing as expected.

This change also removes the support of the unused and not functioning
pragma XFAIL_Nameres.