Fix p_body_part_for_decl for when body part is in separate unit
When the body part of a declaration is in a separate unit,
`p_body_part_for_decl` should check for the next part in the separate
body explicitely since no `__nextpart` symbol exists for it in the
environement.

This check was only done for declarations included in public parts,
that change extends it to private parts.