contrib/highlight: expand all token kind ranges
An upcoming commit in Langkit will change the order of enumeration
values in the Libadalang.Common.Token_Kind type, so the semantic of
ranges will change.

The expanded form is much easier to work with anyway: readers only need
a simple text search to find the highlighting for a given token type.