Libadalang.Project_Provider: fix support for projects w/o Ada sources
We used to raise a Runtime_Missing_Error exception for this edge case:
GPR2 sees no Ada sources in the loaded project, and so does not load its
runtime even though `With_Runtime => True` was passed.

This fixes testsuite failures visible in the upcoming transition to GPR2
for the C/Python/Java APIs.