c_api/gpr_config_file: robustify testcase
The fact that there is no language in the project (due to the content of
empty.cgpr) is an error with GNATCOLL.Projects, but will not be with
GPR2. Rework the testcase so that the presence of the CGPR always
trigger an error.

Do a similar change in the corresponding Java API testcase.