GPR2 project provider: fix handling of empty unit names
While there is a priori no valid use case to call Get_From_Provider with
an empty unit name, the current implementation propagates an assertion
error exception to callers that looks like a LAL bug. Gracefully handle
this case so that it is clearer that the problem is on the caller's
side.