Libadalang.Env_Hooks: run PLE even on unit that could not be parsed
Not doing so when the root node is null is dubious in the first place:
PLE can run on all units, even those that have parsing errors. In
addition, running PLE once keeps track of the fact that PLE was
requested for this unit, and will thus automatically re-run PLE when
this unit will get reparsed, which is necessary to have correct
incremental behavior.

Testcase added in XFAIL mode because there still seems to be an env
cache bug in Langkit.