Data Decomposition: fix loading of JSON with overloaded entities
In the presence of overloaded entities in the Ada source code, it is
possible to get distinct type entries in the JSON with the same fully
qualified name.

To support this use case (which previously aborted the loading of
JSONs), change the repinfo collection internals to:

* allow the storage of multiple entries with the same FQN,
* track the location information associated to each entry,
* use node slocs to disambiguate which entry to return during lookup
  using the entries's location information.

Update testcases which assumed that entries with same FQN would make
loading crash, and testcases which assumed that type entries locations
do not interfere in the lookup mechanism.