Data Decomposition: do not crash when base record repinfo is missing
Do not crash in this case as it may be hard for users to provide the
JSON for the base type's repinfo (for instance
Ada.Finalization.Controlled, which would require to rebuild the
runtime). Just log this information and stop the iteration there: users
just won't have access to repinfo for the base type's components.