gitlab-ci.plan: do not build gnatformat explicitly
Nowadays, GNATformat is part of GNAT, so just having an explicit build
of GNAT with the `--smart-rebuild` anod option and `gnatformat-edge` in
`anod vcs` is enough to trigger its build.

Note that since GNATformat has been stabilized, its build on the safe
track does *not* use the Libadalang wavefront, making the current
duplicated build useless.