CB: do not override GPR2 sources
Nowadays, LibGPR2 does not depend on Libadalang nor Langkit directly
(the dependency on Langkit is vendored), so unconditionally rebuilding
everything that depends on the `eng/gpr/gpr` repository through anod's
`--smart-rebuild` option serves no purpose in the context of our CB, and
takes a lot of time. Do not override GPR2 sources to get rid of this.