Libadalang.Config_Pragmas: rework the API not to force parsing
Data types in this API store analysis units directly.

This means that in order to create the config pragmas mapping for a
given project, all sources of that project + for the config pragmas
files need to be parsed. For cases like opening a project in order to
analyze only a couple of source files, this is very inefficient, both
for processing time and for memory consumption.

This commit reworks the data types to avoid this: store filenames
instead of analysis units.